comments + caster class

+++ render.py
@@ -1,18 +1,36 @@
 from rectangle import *
 
+# Written by Ron Broner 
+# All rights reserved 
+
 class Render():
 
+	# Initialize all variables local to 3D rendering
 	def __init__(self,canvas,winWidth,winHeight,rayCount):
+		# Drawing surface canvas
 		self.canvas = canvas
+
+		# Window dimensions
 		self.winHeight = winHeight
 		self.winWidth = winWidth
+
+		# Number of rays
 		self.rayCount = rayCount
+
+		# Width of each rendered rectangle 
 		self.rectWidth = winWidth/rayCount
+
+		# Litst of rendered rectangles
 		self.renderRects = []
+
+		# Scaler constant to add to y coordinate to simulate moving up/down
 		self.povHeight = 0
+
+		# Initialize the rendered rectangles
 		for i in range(rayCount):
 			self.renderRects.append(Rectangle(canvas,0,0,0,0))
 
+		# Initialize suplementary floor/sky rendered rectangles
 		self.floorRects = []
 		self.skyRects = []
 		for i in range(rayCount):
@@ -21,6 +39,13 @@ class Render():
 			self.floorRects[i].changeColor((0,98,0))
 			self.skyRects[i].changeColor((135,206,235))
 
+
+	# Input: array of 2D rays (which come from the 2D Caster object)
+	#
+	# Goes through all the 2D rays and based on their length (i.e. function of intersection with obstacle)
+	# draws an array of thin vertical rectangles (one per ray) which vary in size and shade to create the illusion of 
+	# 3D depth perception. This method also draws the floor and sky, by basically drawing similar thin rectangles below 
+	# and above the wall rectangles in green adnd blue respectively.
 
 	def updateRays(self,rays):
 		for i in range(len(self.renderRects)):
@@ -41,14 +66,13 @@ class Render():
 
 			# Change rectangle color
 			fFactor = 3
-			c = dist/fFactor
+			c = int(dist/fFactor)
 			if c > 255:
 				c = 255
 			self.renderRects[i].changeColor((c,c,c))
 
 			# Finally draw the fixed rectangles 
 			self.renderRects[i].draw()
-
 
 
 			# Now draw the floor/sky
@@ -61,9 +85,11 @@ class Render():
 			self.skyRects[i].resize(self.rectWidth,yCoord) 
 			self.skyRects[i].draw()
 
+	# Moves the "Camera" POV to simulate tilting your head up/down
 	def moveUpDown(self,amount):
 		self.povHeight = self.povHeight + amount
 
+	# Returns the POV height (the angle of polar head tilt)
 	def getPovHeight(self):
 		return self.povHeight
